move handlesearchaction
